requirements.txt
# Interfaces to system services
redis==2.10.5
psycopg2==2.7.1

# "Scientific" packages
scipy==0.19.0
gensim==2.1.0
nltk==3.2.4
matplotlib==2.0.2

# Generic webdev
Scrapy==1.4.0
Django==1.11.2
djangorestframework==3.6.3
django_dirtyfields==1.2.1
django-extensions==1.7.9

# Testing
factory_boy==2.8.1
fakeredis==0.8.2